requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pandas==1.4.2
numpy==1.22.4
scikit-learn==1.1.2
python-dotenv==1.0.0
